chore: seo setup update
