+++ frontend/package.json
@@ -9,13 +9,15 @@
     "start": "next start",
     "lint": "eslint",
     "deploy": "vercel --prod",
-    "preview": "vercel"
+    "preview": "vercel",
+    "generate-og": "node scripts/generate-og-image.js"
   },
   "dependencies": {
     "clsx": "^2.1.1",
     "ioredis": "^5.8.0",
     "lucide-react": "^0.544.0",
     "next": "15.5.4",
+    "next-seo": "^6.8.0",
     "next-sitemap": "^4.2.3",
     "react": "19.1.0",
     "react-dom": "19.1.0",
@@ -29,6 +31,7 @@
     "@types/react-dom": "^19",
     "eslint": "^9",
     "eslint-config-next": "15.5.4",
+    "sharp": "^0.34.4",
     "tailwindcss": "^4",
     "typescript": "^5"
   }
