+++ frontend/src/app/layout.tsx
@@ -7,10 +7,65 @@ import ScrollToTop from '@/components/ScrollToTop';
 const inter = Inter({ subsets: ['latin'] });
 
 export const metadata: Metadata = {
-  title: '정부 사이트 장애 현황 모니터링',
-  description: '정부 서비스들의 상태를 모니터링하고 분석하는 대시보드 (10분마다 업데이트)',
+  title: {
+    default: '정부서비스 장애 현황',
+    template: '%s | 정부서비스 장애 현황'
+  },
+  description: '정부24, 공공서비스 등 주요 정부 사이트의 접속 오류와 장애 현황을 실시간으로 모니터링합니다. 10분마다 최신 데이터로 업데이트.',
+  keywords: ['정부사이트', '장애현황', '모니터링', '정부24', '공공서비스', '접속오류', '실시간'],
+  authors: [{ name: '김루비', url: 'https://anb-network.com' }],
+  creator: '김루비',
+  publisher: 'A&B Network',
+  formatDetection: {
+    email: false,
+    address: false,
+    telephone: false,
+  },
+  metadataBase: new URL('https://gov-status.vercel.app'),
+  alternates: {
+    canonical: '/',
+  },
+  openGraph: {
+    title: '정부 사이트 장애 현황',
+    description: '정부24, 공공서비스 등 주요 정부 사이트의 접속 오류와 장애 현황을 실시간으로 모니터링합니다. 10분마다 최신 데이터로 업데이트.',
+    url: 'https://gov-status.vercel.app',
+    siteName: '정부 사이트 장애 현황 모니터링',
+    images: [
+      {
+        url: '/og-image.png',
+        width: 1200,
+        height: 630,
+        alt: '정부 사이트 장애 현황 모니터링',
+      },
+    ],
+    locale: 'ko_KR',
+    type: 'website',
+  },
+  twitter: {
+    card: 'summary_large_image',
+    title: '정부 사이트 장애 현황',
+    description: '정부24, 공공서비스 등 주요 정부 사이트의 접속 오류와 장애 현황을 실시간으로 모니터링합니다.',
+    images: ['/og-image.png'],
+    creator: '@anb_network',
+  },
+  robots: {
+    index: true,
+    follow: true,
+    googleBot: {
+      index: true,
+      follow: true,
+      'max-video-preview': -1,
+      'max-image-preview': 'large',
+      'max-snippet': -1,
+    },
+  },
   icons: {
     icon: '/favicon.svg',
+    shortcut: '/favicon.svg',
+    apple: '/favicon.svg',
+  },
+  verification: {
+    google: '4vZm2AHqQ4NzjmZFA5TQz3JXM4NNdrRPLoPHxFxzKUk',
   },
 };
